pyproject.toml
[project]
name = "devstrap"
version = "0.1.0"
description = "Cross-platform CLI tool to install dev tools"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "inquirerpy>=0.3.4",
]

[project.scripts]
devstrap = "devstrap.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black>=26.3.1",
    "isort>=8.0.1",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.15.5",
    "types-pyyaml>=6.0.12.20250915",
    "xenon>=0.9.3",
]
